initial version + refactoring
