+++ .pre-commit-config.yaml
@@ -18,4 +18,4 @@ repos:
     rev: v2.3.1
     hooks:
     -   id: autoflake
-        args: [--remove-all-unused-imports, --in-place]
+        args: [--remove-all-unused-imports, --ignore-init-module-imports, --in-place]
