+++ src/dlomix/__init__.py
@@ -1,10 +1,5 @@
-__version__ = "0.1.9"
+from .config import _BACKEND
 
-META_DATA = {
-    "author": "Wilhelm Lab",
-    "author_email": "o.shouman@tum.de",
-    "description": "Deep Learning for Proteomics",
-    "package_name": "DLOmix",
-    "copyright_text": "2025, Wilhelm Lab, TU Munich, School of Life Sciences",
-    "github_url": "https://github.com/wilhelm-lab/dlomix",
-}
+__all__ = [
+    "_BACKEND",  # Also expose the current backend name
+]
