+++ src/dlomix/_metadata.py
@@ -1,0 +1,8 @@
+__version__ = "0.1.9"
+__author__ = "Wilhelm Lab"
+__author_email__ = "o.shouman@tum.de"
+__license__ = "MIT"
+__description__ = "Deep Learning for Proteomics"
+__package__ = "DLOmix"
+__copyright__ = "2025, Wilhelm Lab, TU Munich, School of Life Sciences"
+__github_url__ = "https://github.com/wilhelm-lab/dlomix"
