+++ src/dlomix/eval/__init__.py
@@ -1,4 +1,14 @@
-from .chargestate import adjusted_mean_absolute_error
-from .rt_eval import timedelta
+from ..config import _BACKEND, PYTORCH_BACKEND, TENSORFLOW_BACKEND
 
-__all__ = ["adjusted_mean_absolute_error", "timedelta"]
+if _BACKEND in TENSORFLOW_BACKEND:
+    from .chargestate import adjusted_mean_absolute_error
+    from .rt_eval import TimeDeltaMetric, timedelta
+elif _BACKEND in PYTORCH_BACKEND:
+    from .chargestate_torch import adjusted_mean_absolute_error
+    from .rt_eval_torch import TimeDeltaMetric, timedelta
+
+__all__ = [
+    "adjusted_mean_absolute_error",
+    "timedelta",
+    "TimeDeltaMetric",
+]
