+++ src/dlomix/layers/__init__.py
@@ -1,11 +1,18 @@
-from .attention import AttentionLayer, DecoderAttentionLayer
-from .attention_torch import AttentionLayerTorch, DecoderAttentionLayerTorch
-from .bi_gru_seq_encoder_torch import BiGRUSequentialEncoder
+from ..config import _BACKEND, PYTORCH_BACKEND, TENSORFLOW_BACKEND
 
-__all__ = [
-    "AttentionLayer",
-    "DecoderAttentionLayer",
-    "AttentionLayerTorch",
-    "DecoderAttentionLayerTorch",
-    "BiGRUSequentialEncoder",
-]
+__all__ = []
+
+if _BACKEND in TENSORFLOW_BACKEND:
+    from .attention import AttentionLayer, DecoderAttentionLayer
+elif _BACKEND in PYTORCH_BACKEND:
+    from .attention import AttentionLayer, DecoderAttentionLayer
+    from .bi_gru_seq_encoder_torch import BiGRUSequentialEncoder
+
+    __all__.append("BiGRUSequentialEncoder")
+
+__all__.extend(
+    [
+        "AttentionLayer",
+        "DecoderAttentionLayer",
+    ]
+)
