+++ src/dlomix/models/__init__.py
@@ -1,24 +1,30 @@
-from .base import *
-from .chargestate import *
-from .chargestate_torch import *
-from .deepLC import *
-from .detectability import *
-from .detectability_torch import *
-from .ionmob_torch import *
-from .prosit import *
-from .prosit_torch import *
+from ..config import _BACKEND, PYTORCH_BACKEND, TENSORFLOW_BACKEND
 
-__all__ = [
-    "RetentionTimePredictor",
-    "PrositRetentionTimePredictor",
-    "DeepLCRetentionTimePredictor",
-    "PrositIntensityPredictor",
-    "ChargeStatePredictor",
-    "DetectabilityModel",
-    "PrositIntensityPredictorTorch",
-    "ChargeStatePredictorTorch",
-    "DetectabilityModelTorch",
-    "Ionmob",
-    "PrositRetentionTimePredictorTorch",
-    "ChargeStatePredictorTorch",
-]
+__all__ = []
+
+if _BACKEND in TENSORFLOW_BACKEND:
+    from .base import RetentionTimePredictor
+    from .chargestate import ChargeStatePredictor
+    from .detectability import DetectabilityModel
+    from .prosit import PrositIntensityPredictor, PrositRetentionTimePredictor
+
+    __all__.append("RetentionTimePredictor")
+
+
+elif _BACKEND in PYTORCH_BACKEND:
+    from .chargestate_torch import ChargeStatePredictor
+    from .detectability_torch import DetectabilityModel
+    from .ionmob_torch import Ionmob
+    from .prosit_torch import PrositIntensityPredictor, PrositRetentionTimePredictor
+
+    __all__.append("Ionmob")
+
+__all__.extend(
+    [
+        "ChargeStatePredictor",
+        "PrositRetentionTimePredictor",
+        "PrositIntensityPredictor",
+        "DetectabilityModel",
+        "ChargeStatePredictor",
+    ]
+)
